Move Product Variant controller into separate trait.
